% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-yoctopuce.R
\name{install_yoctopuce}
\alias{install_yoctopuce}
\title{Install the yoctopuce Python library}
\usage{
install_yoctopuce(..., envname = "r-yoctopuce")
}
\arguments{
\item{...}{additional libraries or named parameters to pass to
`reticulate::py_install`.}

\item{envname}{character The name of a Python environment where to install the
library. Not to be confused with R language environments!}
}
\description{
Python must be installed before installing the library. You can
use `reticulate::install_python` to install Python. Python and the
yoctopuce Python library need to be installed once, when first using R
package 'yoctopuce'.
}
