% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoctopuce.R
\docType{package}
\name{yoctopuce-package}
\alias{yoctopuce}
\alias{yoctopuce-package}
\title{yoctopuce: YoctoPuce USB modules}
\description{
Bridge between R and the yoctopuce Python library, making it possible to control and adquire data from YoctoPuce USB interface modules.
}
\details{
Package 'yoctopuce' provides wrapper functions that call functions
in the 'yoctopuce' Python library through package 'reticulate'.
}
\references{
The documentation of the 'yoctopuce' Python library is installed together
with the library and also available on-line at
\url{https://www.yoctopuce.com/} under the Support menu.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.r4photobiology.info}
  \item \url{https://github.com/aphalo/yoctopuce}
  \item Report bugs at \url{https://github.com/aphalo/yoctopuce/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} (\href{https://orcid.org/0000-0003-3385-972X}{ORCID})

}
