% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y-initialise.R
\name{y_initialise}
\alias{y_initialise}
\alias{y_initialize}
\title{Initialise the API}
\usage{
y_initialise(..., hub.url = "localhost:4444")

y_initialize(..., hub.url = "localhost:4444")
}
\arguments{
\item{...}{character Names of the yoctopuce APIs to import.}

\item{hub.url}{character URL including port of the hub to connect to,
defaults to a virtual hub running locally, but hardware and virtual
hub can be accessed through a LAN or the internet.}
}
\description{
Import the API and module-specific extensions and register to communicate
through a virtual or physical hub.
}
\details{
With this function we initialize the API and register to use a specific
hub. The main API can be accessed through object \code{yocto_api} using
the \code{$} notation. Be aware that there is more than one level of
nesting. Say for 'yoctopuce.yocto_relay' the object \code{yocto_relay} is
created and can be used to access its members.

This function uses the Python API to check that the requested API extensions
are available, so it should work unchanged with future updates to the
YoctoPuce Python library, including after new modules are released as long
as the Python libray installed is up-to-date.
}
